sparse_word_matrix.R
This function converts a text data file into a sparse word matrix.

The input file is described as follows: rows = text units (tweets, documents, etc.), and one column contains the text to be analyzed. 

The output file is a (t x w) sparse matrix where t is the number of participants, and w is the number of words of interest.

A single cell within the matrix lets us know whether a given word is present for a given text unit (document, tweet, etc.).